[#1]refactor(dto): DTO 패키지 삭제 및 domain 패키지로 이동
- Create, Read 객체들도 Model로 간주
- 이런 model들은 외부에서 직접 의존해도 괜찮다고 판단
