+++ src/main/java/baksakcci/shoppingmall/order/application/port/PlaceOrderService.java
@@ -1,6 +1,6 @@
 package baksakcci.shoppingmall.order.application.port;
 
-import baksakcci.shoppingmall.order.application.dto.OrderCreate;
+import baksakcci.shoppingmall.order.domain.OrderCreate;
 import baksakcci.shoppingmall.order.domain.Order;
 
 public interface PlaceOrderService {
