+++ src/main/java/baksakcci/shoppingmall/order/domain/OrderCreate.java
@@ -1,4 +1,4 @@
-package baksakcci.shoppingmall.order.application.dto;
+package baksakcci.shoppingmall.order.domain;
 
 import java.util.List;
 import lombok.Builder;
