+++ src/main/java/baksakcci/shoppingmall/order/presentation/OrderController.java
@@ -2,7 +2,7 @@ package baksakcci.shoppingmall.order.presentation;
 
 import baksakcci.shoppingmall.order.domain.Order;
 import baksakcci.shoppingmall.order.application.port.PlaceOrderService;
-import baksakcci.shoppingmall.order.presentation.request.OrderRequest;
+import baksakcci.shoppingmall.order.domain.OrderCreate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,7 +17,7 @@ public class OrderController {
     private final PlaceOrderService placeOrderService;
 
     @PostMapping("/create")
-    public Order place(@RequestBody OrderRequest orderRequest) {
-        return placeOrderService.placeOrder(OrderRequest.toDto(1L, orderRequest));
+    public Order place(@RequestBody OrderCreate orderCreate) {
+        return placeOrderService.placeOrder(orderCreate);
     }
 }
