--- src/main/java/baksakcci/shoppingmall/order/presentation/response/OrderResponse.java
@@ -1,10 +1,0 @@
-package baksakcci.shoppingmall.order.presentation.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class OrderResponse {
-    private Long orderId;
-}
